requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
zai-sdk>=0.0.2
cohere>=5.0.0
pinecone>=7.0.0
pydantic>=2.0.0
firebase-admin>=6.6.0
